Make grid layout reusable. Change the file structure.
