+++ NextJS/components/FontProject/charIdx.js
