+++ NextJS/components/GridLayout/Grid.js
@@ -1,0 +1,11 @@
+export default function Grid(props) {
+
+    return (
+        <div className='flex flex-wrap w-100'>
+            {props.children}
+        </div>
+    )
+
+
+
+}
