+++ NextJS/pages/index.js
@@ -1,8 +1,6 @@
 import Head from 'next/head'
 
-import DownloadForm from '../components/DownloadForm.js'
-import Pagination from '../components/Pagination'
-import ProjectDetail from '../components/ProjectDetail.js'
+import ItemsGrid from '../components//GridLayout/ItemsGrid'
 import FontProjectInterface from '../components/FontProjectInterface.js'
 
 
@@ -17,27 +15,9 @@ export default function Home() {
       <main id='root'>
         <h1>Font Project</h1>
         <FontProjectInterface />
-        <ProjectDetail />
+        <ItemsGrid col='3' row='4' count='95' type='char' />
       </main>
 
-      {/* 
-      
-      If using tachyons, no need to have an extra style. 
-
-      <style jsx global>{`
-        html,
-        body {
-          padding: 0;
-          margin: 0;
-          font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto,
-            Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,
-            sans-serif;
-        }
-
-        * {
-          box-sizing: border-box;
-        }
-      `}</style> */}
     </div>
   )
 }
